+++ src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/machine/KJSWrappingMultiblockBuilder.java
@@ -1,7 +1,14 @@
 package com.gregtechceu.gtceu.integration.kjs.builders.machine;
 
+import com.gregtechceu.gtceu.api.block.MetaMachineBlock;
+import com.gregtechceu.gtceu.api.blockentity.MetaMachineBlockEntity;
+import com.gregtechceu.gtceu.api.item.MetaMachineItem;
 import com.gregtechceu.gtceu.api.machine.MultiblockMachineDefinition;
+import com.gregtechceu.gtceu.api.machine.multiblock.MultiblockControllerMachine;
+import com.gregtechceu.gtceu.api.machine.multiblock.WorkableElectricMultiblockMachine;
 import com.gregtechceu.gtceu.api.registry.registrate.BuilderBase;
+import com.gregtechceu.gtceu.api.registry.registrate.MultiblockMachineBuilder;
+import com.gregtechceu.gtceu.common.registry.GTRegistration;
 
 import net.minecraft.resources.ResourceLocation;
 
@@ -47,5 +54,22 @@ public class KJSWrappingMultiblockBuilder extends BuilderBase<MultiblockMachineD
         // should never happen.
         throw new IllegalStateException("Empty tiered multiblock builder " + Arrays.toString(tieredBuilder.get()) +
                 " With id " + tieredBuilder.id);
+    }
+
+    public static MultiblockMachineBuilder createKJSMulti(ResourceLocation id) {
+        return MultiblockMachineBuilder.createMulti(GTRegistration.REGISTRATE, id.getPath(),
+                WorkableElectricMultiblockMachine::new,
+                MetaMachineBlock::new,
+                MetaMachineItem::new,
+                MetaMachineBlockEntity::createBlockEntity);
+    }
+
+    public static MultiblockMachineBuilder createKJSMulti(ResourceLocation id,
+                                                          KJSTieredMachineBuilder.CreationFunction<? extends MultiblockControllerMachine> machine) {
+        return MultiblockMachineBuilder.createMulti(GTRegistration.REGISTRATE, id.getPath(),
+                machine::create,
+                MetaMachineBlock::new,
+                MetaMachineItem::new,
+                MetaMachineBlockEntity::createBlockEntity);
     }
 }
